move migrations validation and conversion logic to field handlers (task #2042)
